refactor: 💡 outsource postcodeSuggestion component from askerPr
